Issue #2882717 by Wim Leers, dawehner: EntityResourceTestBase: assert that PATCHed and POSTed entities contain the intended values

+++ core/modules/rest/tests/src/Functional/EntityResource/Comment/CommentResourceTestBase.php
@@ -219,7 +219,7 @@ abstract class CommentResourceTestBase extends EntityResourceTestBase {
       ],
       'entity_id' => [
         [
-          'target_id' => EntityTest::load(1)->id(),
+          'target_id' => (int) EntityTest::load(1)->id(),
         ],
       ],
       'field_name' => [
